# docs/api/routes.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Titan — inventario de rotas",
    "version": "1.0.0",
    "description": "Mapa da base fde43f8. Schemas de payload pendentes estão marcados por operação; não é um contrato completo de negócios."
  },
  "paths": {
    "/local/v1/access/audit": {
      "get": {
        "operationId": "local_GET__local_v1_access_audit",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./access-policy.openapi.json"
        }
      }
    },
    "/local/v1/access/policy": {
      "get": {
        "operationId": "local_GET__local_v1_access_policy",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./access-policy.openapi.json"
        }
      },
      "post": {
        "operationId": "local_POST__local_v1_access_policy",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./access-policy.openapi.json"
        }
      }
    },
    "/local/v1/account/password": {
      "post": {
        "operationId": "local_POST__local_v1_account_password",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./account-security.openapi.json"
        }
      }
    },
    "/local/v1/account/recover": {
      "post": {
        "operationId": "local_POST__local_v1_account_recover",
        "security": [],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./account-security.openapi.json"
        }
      }
    },
    "/local/v1/account/recovery/revoke": {
      "post": {
        "operationId": "local_POST__local_v1_account_recovery_revoke",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./account-security.openapi.json"
        }
      }
    },
    "/local/v1/account/sessions": {
      "get": {
        "operationId": "local_GET__local_v1_account_sessions",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./account-security.openapi.json"
        }
      }
    },
    "/local/v1/account/sessions/revoke-others": {
      "post": {
        "operationId": "local_POST__local_v1_account_sessions_revoke_others",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./account-security.openapi.json"
        }
      }
    },
    "/local/v1/capabilities": {
      "get": {
        "operationId": "local_GET__local_v1_capabilities",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "pending",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented"
      }
    },
...
    "/local/v1/login": {
      "post": {
        "operationId": "local_POST__local_v1_login",
        "security": [],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "pending",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented"
      }
    },
    "/local/v1/logout": {
      "post": {
        "operationId": "local_POST__local_v1_logout",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "pending",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented"
      }
    },
    "/local/v1/me": {
      "get": {
        "operationId": "local_GET__local_v1_me",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "pending",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented"
      }
    },
    "/local/v1/module-contracts": {
      "post": {
        "operationId": "local_POST__local_v1_module_contracts",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "pending",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented"
      }
    },
...
    "/local/v1/staff": {
      "get": {
        "operationId": "local_GET__local_v1_staff",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "pending",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented"
      },
      "post": {
        "operationId": "local_POST__local_v1_staff",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "pending",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented"
      }
    },
    "/local/v1/staff/{id}/password-reset": {
      "post": {
        "operationId": "local_POST__local_v1_staff__id_password_reset",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./account-security.openapi.json"
        }
      }
    },
    "/local/v1/staff/{id}/sessions/revoke": {
      "post": {
        "operationId": "local_POST__local_v1_staff__id_sessions_revoke",
        "security": [
          {
            "BearerSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "./http-errors.openapi.json#/components/parameters/ErrorFormat"
          }
        ],
        "responses": {
          "default": {
            "description": "Consultar contrato detalhado do dominio; formato legado ou erro negociado conforme header."
          }
        },
        "x-payload-contract": "documented",
        "x-source": "backend/internal/localapi/server.go",
        "x-state": "implemented",
        "externalDocs": {
          "description": "Contrato detalhado desta operação",
          "url": "./account-security.openapi.json"
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerSession": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}